+++ Tests/RichTextKitTests/RichTextViewRepresentable+ColorTests.swift
@@ -1,30 +1,87 @@
 //
-//  RichTextViewRepresentable+AlignmentTests.swift
+//  RichTextViewRepresentable+ColorTests.swift
 //  RichTextKitTests
 //
-//  Created by Daniel Saidi on 2021-12-30.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//  Created by Daniel Saidi on 2022-12-06.
+//  Copyright © 2022 Daniel Saidi. All rights reserved.
 //
 
+#if canImport(UIKit)
+import UIKit
+#endif
+
+#if os(macOS)
+import AppKit
+#endif
+
+#if os(iOS) || os(macOS) || os(tvOS)
 import RichTextKit
 import XCTest
 
-class RichTextViewRepresentable_ColorAlignmentTests: XCTestCase {
+class RichTextViewRepresentable_ColorTests: XCTestCase {
 
-    let string = NSMutableAttributedString(string: "foo bar baz")
+    var textView: RichTextViewRepresentable!
 
-    func testSettingColorWithinRangeOnlySetsAttributeWithinTheProvidedRange() {
-        let color = ColorRepresentable.yellow
-        let range = NSRange(location: 4, length: 3)
-        let noRange = NSRange(location: 0, length: 0)
-        string.setRichTextAttribute(.foregroundColor, to: color, at: range)
+    let color = ColorRepresentable.red
+    let noRange = NSRange(location: 0, length: 0)
+    let selectedRange = NSRange(location: 4, length: 3)
 
-        let attr1: ColorRepresentable? = string.richTextAttribute(.foregroundColor, at: range)
-        XCTAssertEqual(attr1, color)
-        let attr2: ColorRepresentable? = string.richTextAttribute(.foregroundColor, at: noRange)
-        XCTAssertNil(attr2)
+    override func setUp() {
+        textView = RichTextView()
+        textView.setup(
+            with: NSAttributedString(string: "foo bar baz"),
+            format: .rtf
+        )
+    }
 
-        XCTAssertEqual(string.richTextAttributes(at: range)[.foregroundColor] as? ColorRepresentable, color)
-        XCTAssertNil(string.richTextAttributes(at: noRange)[.foregroundColor] as? ColorRepresentable)
+    func assertEqualColor(_ attr: Any?) {
+        XCTAssertEqual(attr as? ColorRepresentable, color)
+    }
+
+    func assertNonEqualColor(_ attr: Any?) {
+        XCTAssertNotEqual(attr as? ColorRepresentable, color)
+    }
+
+
+    func testCurrentBackgroundColorWorksForSelectedRange() {
+        textView.setSelectedRange(selectedRange)
+        textView.setCurrentBackgroundColor(to: color)
+        XCTAssertEqual(textView.currentBackgroundColor, color)
+        assertEqualColor(textView.currentRichTextAttributes[.backgroundColor])
+        assertEqualColor(textView.richTextAttributes(at: selectedRange)[.backgroundColor])
+        assertEqualColor(textView.typingAttributes[.backgroundColor])
+    }
+
+    func testCurrentBackgroundColorWorksForNoSelectedRange() {
+        textView.setSelectedRange(noRange)
+        textView.setCurrentBackgroundColor(to: color)
+        #if os(iOS) || os(tvOS)
+        XCTAssertEqual(textView.currentBackgroundColor, color)
+        assertEqualColor(textView.currentRichTextAttributes[.backgroundColor])
+        #endif
+        XCTAssertNil(textView.richTextAttributes(at: selectedRange)[.backgroundColor])
+        assertEqualColor(textView.typingAttributes[.backgroundColor])
+    }
+
+
+    func testCurrentForegroundColorWorksForSelectedRange() {
+        textView.setSelectedRange(selectedRange)
+        textView.setCurrentForegroundColor(to: color)
+        XCTAssertEqual(textView.currentForegroundColor, color)
+        assertEqualColor(textView.currentRichTextAttributes[.foregroundColor])
+        assertEqualColor(textView.richTextAttributes(at: selectedRange)[.foregroundColor])
+        assertEqualColor(textView.typingAttributes[.foregroundColor])
+    }
+
+    func testCurrentForegroundColorWorksForNoSelectedRange() {
+        textView.setSelectedRange(noRange)
+        textView.setCurrentForegroundColor(to: color)
+        #if os(iOS) || os(tvOS)
+        XCTAssertEqual(textView.currentForegroundColor, color)
+        assertEqualColor(textView.currentRichTextAttributes[.foregroundColor])
+        #endif
+        XCTAssertNotEqual(textView.richTextAttributes(at: selectedRange)[.foregroundColor] as? ColorRepresentable, color)
+        assertEqualColor(textView.typingAttributes[.foregroundColor])
     }
 }
+#endif